Swap and Select
Added the swap and selectBubble methods to BubbleMatrix. In order to do
this I also added a copy contructor to BubbleButton as well as a select
boolean value to the Bubble class with its getter and setter methods.
The select boolean value is initial set to false, and the setter
methods negates selects current value and returns it.